Relocate SaveChanges to own file
