+++ packages/airview-cms/src/features/cms/toolbar/toolbar.js
@@ -11,20 +11,13 @@ import {
   clearMetaDataEdits,
   selectMetaEditorEnabledStatus,
   selectDoesMetaEditorHaveEdits,
-  selectMetaEditorData,
 } from "../meta-editor";
-import {
-  disableCms,
-  selectCmsContext,
-  selectCmsBusyStatus,
-} from "../cms.slice";
+import { disableCms, selectCmsBusyStatus } from "../cms.slice";
 import {
   enableCreatePullRequestModal,
   selectCanCreatePullRequest,
 } from "../create-pull-request";
-import { usePutEntryMutation } from "../../store";
-import { selectWorkingBranch } from "./working-branch.slice";
-import matter from "gray-matter";
+import { SaveChanges } from "./save-changes";
 
 export function ToolBar() {
   const dispatch = useDispatch();
@@ -203,32 +196,5 @@ export function ToolBar() {
         </Box>
       </Toolbar>
     </AppBar>
-  );
-}
-
-function SaveChanges() {
-  const hasEdits = useSelector(selectDoesMetaEditorHaveEdits);
-  const edits = useSelector(selectMetaEditorData);
-  const [putEntry, { isLoading }] = usePutEntryMutation();
-  const id = useSelector(selectCmsContext);
-  const branch = useSelector(selectWorkingBranch);
-
-  const handleOnClick = () => {
-    const data = {
-      _index: btoa(matter.stringify("", edits)),
-    };
-
-    putEntry({ id, branch, data });
-  };
-
-  return (
-    <Button
-      variant="text"
-      size="small"
-      disabled={!hasEdits || isLoading}
-      onClick={handleOnClick}
-    >
-      Save Changes
-    </Button>
   );
 }
